BUILD-SES-001: align accept body with OQ-039 serving multiplier.
Optional int servingMultiplier (default 1, range 1–12); test omitted field;
WORK_QUEUE marks AUTH/SES awaiting_human_review.

+++ backend/TrueSight.Api/Features/Sessions/AcceptRecipe/Command.cs
@@ -2,5 +2,5 @@ using MediatR;
 
 namespace TrueSight.Api.Features.Sessions.AcceptRecipe;
 
-public sealed record AcceptRecipeCommand(string RecipeId, decimal ServingMultiplier) : IRequest<RecipeSessionResponse>;
+public sealed record AcceptRecipeCommand(string RecipeId, int ServingMultiplier) : IRequest<RecipeSessionResponse>;
